Convert to bazel, break out dependencies.
