+++ .bazelrc
@@ -1,0 +1,1 @@
+build --copt=-std=c++17
