+++ player.py
@@ -1,0 +1,25 @@
+# Copyright 2021 Mike Godwin
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+class Player(object):
+
+  def __init__(self, board, score):
+    self.board = board
+    self.score = score
+    self.streams = []
+    self.meadows = []
+
+  def total(self):
+    return sum([v for v in self.score.values()])
